+++ hashes/index.js
@@ -1,5 +1,9 @@
 const idForBlock = require('./id_for_block');
-const noLocktimeIdForTransaction = require('./no_locktime_id_for_transaction');
+const idForTransactionComponents = require('./id_for_transaction_components');
 const previousBlockId = require('./previous_block_id');
 
-module.exports = {idForBlock, noLocktimeIdForTransaction, previousBlockId};
+module.exports = {
+  idForBlock,
+  idForTransactionComponents,
+  previousBlockId,
+};