+++ index.js
@@ -1,9 +1,11 @@
 const {compactIntAsNumber} = require('./numbers');
 const {componentsOfTransaction} = require('./transactions');
 const {idForBlock} = require('./hashes');
-const {noLocktimeIdForTransaction} = require('./hashes');
+const {idForTransactionComponents} = require('./hashes');
+const {noLocktimeIdForTransaction} = require('./transactions');
 const {numberAsCompactInt} = require('./numbers');
 const {previousBlockId} = require('./hashes');
+const {queryTransactions} = require('./transactions');
 const {scriptAsScriptElements} = require('./script');
 const {scriptElementsAsScript} = require('./script');
 
@@ -11,9 +13,11 @@ module.exports = {
   compactIntAsNumber,
   componentsOfTransaction,
   idForBlock,
+  idForTransactionComponents,
   noLocktimeIdForTransaction,
   numberAsCompactInt,
   previousBlockId,
+  queryTransactions,
   scriptAsScriptElements,
   scriptElementsAsScript,
 };