+++ transactions/index.js
@@ -1,4 +1,11 @@
 const componentsOfTransaction = require('./components_of_transaction');
+const noLocktimeIdForTransaction = require('./no_locktime_id_for_transaction');
 const parseTransaction = require('./parse_transaction');
+const queryTransactions = require('./query_transactions');
 
-module.exports = {componentsOfTransaction, parseTransaction};
+module.exports = {
+  componentsOfTransaction,
+  noLocktimeIdForTransaction,
+  parseTransaction,
+  queryTransactions,
+};